+++ record_android/android/src/main/kotlin/com/llfbandit/record/methodcall/MethodCallHandlerImpl.kt
@@ -115,7 +115,7 @@ class MethodCallHandlerImpl(
     }
 
     private fun createRecorder(recorderId: String) {
-        val recorder = RecorderWrapper(recorderId, messenger)
+        val recorder = RecorderWrapper(appContext, recorderId, messenger)
         recorder.setActivity(activity)
         recorders[recorderId] = recorder
 
@@ -136,11 +136,7 @@ class MethodCallHandlerImpl(
     }
 
     private fun getRecordConfig(call: MethodCall): RecordConfig {
-        val device = if (Build.VERSION.SDK_INT >= 23) {
-            DeviceUtils.deviceInfoFromMap(appContext, call.argument("device"))
-        } else {
-            null
-        }
+        val androidConfig = call.argument("androidConfig") as Map<*, *>?
 
         return RecordConfig(
             call.argument("path"),
@@ -148,10 +144,15 @@ class MethodCallHandlerImpl(
             Utils.firstNonNull(call.argument("bitRate"), 128000),
             Utils.firstNonNull(call.argument("sampleRate"), 44100),
             Utils.firstNonNull(call.argument("numChannels"), 2),
-            device,
+            if (Build.VERSION.SDK_INT >= 23) {
+                DeviceUtils.deviceInfoFromMap(appContext, call.argument("device"))
+            } else {
+                null
+            },
             Utils.firstNonNull(call.argument("autoGain"), false),
             Utils.firstNonNull(call.argument("echoCancel"), false),
-            Utils.firstNonNull(call.argument("noiseSuppress"), false)
+            Utils.firstNonNull(call.argument("noiseSuppress"), false),
+            Utils.firstNonNull(androidConfig?.get("useLegacy") as Boolean?, false),
         )
     }
 }