+++ record_android/android/src/main/kotlin/com/llfbandit/record/methodcall/RecorderWrapper.kt
@@ -1,16 +1,23 @@
 package com.llfbandit.record.methodcall
 
 import android.app.Activity
-import com.llfbandit.record.record.AudioRecorder
+import android.content.Context
+import com.llfbandit.record.record.recorder.AudioRecorder
 import com.llfbandit.record.record.RecordConfig
 import com.llfbandit.record.record.bluetooth.BluetoothScoListener
+import com.llfbandit.record.record.recorder.IRecorder
+import com.llfbandit.record.record.recorder.MediaRecorder
 import com.llfbandit.record.record.stream.RecorderRecordStreamHandler
 import com.llfbandit.record.record.stream.RecorderStateStreamHandler
 import io.flutter.plugin.common.BinaryMessenger
 import io.flutter.plugin.common.EventChannel
 import io.flutter.plugin.common.MethodChannel
 
-internal class RecorderWrapper(recorderId: String, messenger: BinaryMessenger): BluetoothScoListener {
+internal class RecorderWrapper(
+    private val context: Context,
+    recorderId: String,
+    messenger: BinaryMessenger
+): BluetoothScoListener {
     companion object {
         const val EVENTS_STATE_CHANNEL = "com.llfbandit.record/events/"
         const val EVENTS_RECORD_CHANNEL = "com.llfbandit.record/eventsRecord/"
@@ -20,7 +27,7 @@ internal class RecorderWrapper(recorderId: String, messenger: BinaryMessenger): 
     private val recorderStateStreamHandler = RecorderStateStreamHandler()
     private var eventRecordChannel: EventChannel?
     private val recorderRecordStreamHandler = RecorderRecordStreamHandler()
-    private var recorder: AudioRecorder? = null
+    private var recorder: IRecorder? = null
 
     init {
         eventChannel = EventChannel(messenger, EVENTS_STATE_CHANNEL + recorderId)
@@ -39,6 +46,9 @@ internal class RecorderWrapper(recorderId: String, messenger: BinaryMessenger): 
     }
 
     fun startRecordingToStream(config: RecordConfig, result: MethodChannel.Result) {
+        if (config.useLegacy) {
+            throw Exception("Unsupported feature from legacy recorder.")
+        }
         startRecording(config, result)
     }
 
@@ -119,7 +129,7 @@ internal class RecorderWrapper(recorderId: String, messenger: BinaryMessenger): 
     private fun startRecording(config: RecordConfig, result: MethodChannel.Result) {
         try {
             if (recorder == null) {
-                recorder = createRecorder()
+                recorder = createRecorder(config)
                 start(config, result)
             } else if (recorder!!.isRecording) {
                 recorder!!.stop(fun(_) = start(config, result))
@@ -131,7 +141,11 @@ internal class RecorderWrapper(recorderId: String, messenger: BinaryMessenger): 
         }
     }
 
-    private fun createRecorder(): AudioRecorder {
+    private fun createRecorder(config: RecordConfig): IRecorder {
+        if (config.useLegacy) {
+            return MediaRecorder(context, recorderStateStreamHandler)
+        }
+
         return AudioRecorder(
             recorderStateStreamHandler,
             recorderRecordStreamHandler