+++ record_android/android/src/main/kotlin/com/llfbandit/record/record/format/Format.kt
@@ -1,7 +1,10 @@
 package com.llfbandit.record.record.format
 
+import android.media.MediaCodecInfo
+import android.media.MediaCodecList
 import android.media.MediaFormat
 import android.util.Log
+import android.util.Range
 import com.llfbandit.record.record.RecordConfig
 import com.llfbandit.record.record.container.IContainerWriter
 import com.llfbandit.record.record.encoder.EncoderListener
@@ -25,21 +28,43 @@ sealed class Format {
      */
     abstract fun getMediaFormat(config: RecordConfig): MediaFormat
 
+    protected open fun adjustSampleRate(format: MediaFormat, sampleRate: Int) {
+        format.setInteger(MediaFormat.KEY_SAMPLE_RATE, sampleRate)
+    }
+
+    private fun adjustBitRate(format: MediaFormat, bitRate: Int) {
+        format.setInteger(MediaFormat.KEY_BIT_RATE, bitRate)
+    }
+
+    protected open fun adjustNumChannels(format: MediaFormat, numChannels: Int) {
+        format.setInteger(MediaFormat.KEY_CHANNEL_MASK, numChannels)
+    }
+
     /**
      * Create an encoder that produces [MediaFormat] output.
      */
     fun getEncoder(
         config: RecordConfig,
         listener: EncoderListener
-    ): IEncoder {
+    ): Pair<IEncoder, MediaFormat> {
 
         val mediaFormat = getMediaFormat(config)
-        val container = getContainer(config.path)
 
         return if (passthrough) {
-            PassthroughEncoder(mediaFormat, listener, container)
+            Pair(
+                PassthroughEncoder(mediaFormat, listener, getContainer(config.path)),
+                mediaFormat
+            )
         } else {
-            MediaCodecEncoder(mediaFormat, listener, container)
+            val encoder = findCodecForAdjustedFormat(config, mediaFormat)
+            encoder ?: throw Exception(
+                "No encoder found for given config $mediaFormat. You should try with other values."
+            )
+
+            Pair(
+                MediaCodecEncoder(encoder, mediaFormat, listener, getContainer(config.path)),
+                mediaFormat
+            )
         }
     }
 
@@ -63,11 +88,77 @@ sealed class Format {
         }
 
         if (value != values[idx]) {
-            Log.d(TAG, "Available values: $values")
-            Log.d(TAG, "Adjusted to: $value")
+            Log.d(TAG, "Available values: ${values.indices.map { values[it] }}")
+            Log.d(TAG, "Adjusted to: ${values[idx]}")
         }
 
         return values[idx]
+    }
+
+    private fun checkBounds(range: Range<Int>, value: Int): Int {
+        if (range.lower > value) {
+            return range.lower
+        } else if (range.upper < value) {
+            return range.upper
+        }
+        return value
+    }
+
+    private fun adjustFormat(
+        caps: MediaCodecInfo.CodecCapabilities,
+        config: RecordConfig,
+        mediaFormat: MediaFormat
+    ): Boolean {
+        if (!caps.isFormatSupported(mediaFormat)) {
+            adjustBitRate(
+                mediaFormat,
+                checkBounds(caps.audioCapabilities.bitrateRange, config.bitRate)
+            )
+            if (caps.audioCapabilities.supportedSampleRates != null) {
+                adjustSampleRate(
+                    mediaFormat,
+                    nearestValue(
+                        caps.audioCapabilities.supportedSampleRates,
+                        config.sampleRate
+                    )
+                )
+            }
+            adjustNumChannels(
+                mediaFormat,
+                nearestValue(
+                    intArrayOf(1, caps.audioCapabilities.maxInputChannelCount),
+                    config.numChannels
+                )
+            )
+
+            return caps.isFormatSupported(mediaFormat)
+        }
+
+        return true
+    }
+
+    private fun findCodecForAdjustedFormat(
+        config: RecordConfig,
+        mediaFormat: MediaFormat
+    ): String? {
+        val codecs = MediaCodecList(MediaCodecList.REGULAR_CODECS)
+
+        for (info in codecs.codecInfos) {
+            if (!info.isEncoder) {
+                continue
+            }
+
+            try {
+                val caps = info.getCapabilitiesForType(mimeTypeAudio)
+                if (caps != null && adjustFormat(caps, config, mediaFormat)) {
+                    return info.name
+                }
+            } catch (e: IllegalArgumentException) {
+                // type is not supported
+            }
+        }
+
+        return null
     }
 
     companion object {